Improved script to notify user if pk3 files are not found
The script will now notify user if no pk3 files are found inside the folder with the mod script. The script will now also work with pk3 files that has "spaces" in the filename.

+++ Addons/install-ecwolf-mod.sh
@@ -2,33 +2,29 @@
 
 # -- find .pk3 files and move into wolf3d roms folder -- #
 for i in *.pk3; do
+   [[ -f "$i" ]] || continue
    mv "$i" "$HOME/RetroPie/roms/ports/wolf3d"
-   filenames="$filenames $i"
+   filenames="\"$i\" $filenames" ## Should fix globbing and prints output in quotes
 done
+[[ -n "$filenames" ]] || { echo "Error: No additional pk3-files found in $PWD"; sleep 10s; exit; }
 
 # -- read user input for filename change -- #
 echo What should we call the mod?
 read filename
-echo renamed mod to $filename
+echo "Renamed mod to $filename"
 
 # -- create .ecwolf file from users input -- #
-cat > "$filename.ecwolf" << EOF
+cat > "$HOME/RetroPie/roms/ports/wolf3d/$filename.ecwolf" << EOF
 cd "."
 "/opt/retropie/ports/ecwolf/ecwolf" --data WL6 --file $filenames
-wait $!
+wait \$!
 EOF
 
-# -- move .ecwolf file into wolf3d roms folder -- #
-mv "$filename.ecwolf" "$HOME/RetroPie/roms/ports/wolf3d"
-
 # -- create runcommand.sh file from users input -- #
-cat > "$filename.sh" << EOF
+cat > "$HOME/RetroPie/roms/ports/$filename.sh" << EOF
 #!/bin/bash
 "/opt/retropie/supplementary/runcommand/runcommand.sh" 0 _PORT_ "ecwolf" "$HOME/RetroPie/roms/ports/wolf3d/$filename.ecwolf"
 EOF
-
-# -- move runcommand.sh file into ports folder -- #
-mv "$filename.sh" "$HOME/RetroPie/roms/ports/$filename.sh"
 
 # -- installation message -- #
 echo "$filename mod installed for ECWolf. Press enter button to continue."
